✨ feat : api fetch 함수, 비밀번호 유효성 검사 등 공통 util 함수 작성
