+++ client/src/API_TEST.tsx
@@ -1,39 +1,11 @@
 import { useEffect, useState } from 'react';
-import axios from 'axios';
 import S_Container from './components/UI/S_Container';
 import { clubType } from './pages/club/club/CreateClub';
-
-//? return 값 res.data 말고, res 로 하면 타입 지정 어떻게 해야할지?
-export const getFetch = async (url: string) => {
-  try {
-    const res = await axios.get(url);
-    return res.data;
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-export const postFetch = async (url: string, newData: any) => {
-  try {
-    const res = await axios.post(url, JSON.stringify(newData), {
-      headers: {
-        'Content-Type': 'application/json',
-        withCredentials: true
-      }
-    });
-
-    // ? res.ok 왜 안 될까?
-    if (res.status === 200 || res.status === 201) {
-      return res;
-    }
-  } catch (err) {
-    console.error(err);
-  }
-};
+import { getFetch, postFetch } from './util/api';
 
 function API_TEST() {
-  const POST_URL = 'https://dev.somojeon.site/clubs';
-  const GET_URL = 'https://dev.somojeon.site';
+  const POST_URL = `${process.env.REACT_APP_URL}/clubs`;
+  const GET_URL = `${process.env.REACT_APP_URL}`;
 
   const [dataFromServer, setDataFromServer] = useState('');
