+++ client/src/util/alertPreparingService.ts
@@ -1,0 +1,5 @@
+const alertPreparingService = () => {
+  alert('준비 중인 서비스입니다.');
+};
+
+export default alertPreparingService;
